Makefile changes for Amazon Linux support
The default Cloud9 environment on AWS spins up on an Amazon Linux image that doesn't have the right libc++ version available for proper clang++ support. As a result, for the time being we can use g++ and its libstdc++ implementation. (Also specifying c++14 as the standard instead of the deprecated c++1y synonym.)